jeorg-spring-5/jeorg-spring-flash/jeorg-spring-flash-set-4/jeorg-spring-flash-4-16/src/test/java/org/jesperancinha/std/flash416/mock/mvc/VampireControllerTest.java
package org.jesperancinha.std.flash416.mock.mvc;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(SpringExtension.class)
@ContextConfiguration(classes = VampireController.class)
class VampireControllerTest {

    @Autowired
    private VampireController vampireController;

    private MockMvc mockMvc;

    @BeforeEach
    void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(vampireController).build();
    }

    @AfterEach
    void tearDown() {

    }

    @Test
    void testArms_whenRequest_thenReturnOk() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.get("/arms"))
                .andExpect(status().isOk())
                .andExpect(content().contentTypeCompatibleWith("text/plain"))
                .andExpect(content().string("Come into these arms again"));
    }

    @Test
    void testBody_whenRequest_thenReturnOk() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.get("/body"))
                .andExpect(status().isOk())
                .andExpect(content().contentTypeCompatibleWith("text/plain"))
                .andExpect(content().string("Lay your body down"));
    }
}